refactor(#741): refactor caverController
